MinDepthOfBinaryTree (Leetcode - Easy solved and learned)
